Move python files under api_tests folder
