+++ api_tests/api_tests.py
@@ -5,33 +5,33 @@ from datetime import datetime as dt
 from datetime import timedelta as td
 import san
 from san.error import SanError
-from api_helper import get_available_metrics_and_queries, \
-                       get_timeseries_metric_data, \
-                       get_histogram_metric_data, \
-                       get_query_data, \
-                       get_marketcap_batch, \
-                       get_min_interval, \
-                       build_histogram_gql_string, \
-                       build_query_gql_string, \
-                       build_timeseries_gql_string
-from html_reporter import generate_html_from_json
-from queries import special_queries
-from discord_bot import send_metric_alert
-from slugs import legacy_asset_slugs
-from json_processor import create_stable_json
-from metric_report import MetricReport
-from slug_report import SlugReport
-from file_utils import save_json_to_file
-from s3 import upload_to_s3
-from config import Config
-from constants import DATETIME_PATTERN_METRIC, \
-                      HISTOGRAM_METRICS_LIMIT, \
-                      BATCH_SIZE, \
-                      METRICS_WITH_LONGER_DELAY, \
-                      METRICS_WITH_ALLOWED_NEGATIVES, \
-                      INTERVAL_TIMEDELTA, \
-                      ERRORS_IN_ROW, \
-                      PYTHON_ENV
+from .api_helper import get_available_metrics_and_queries, \
+                        get_timeseries_metric_data, \
+                        get_histogram_metric_data, \
+                        get_query_data, \
+                        get_marketcap_batch, \
+                        get_min_interval, \
+                        build_histogram_gql_string, \
+                        build_query_gql_string, \
+                        build_timeseries_gql_string
+from .html_reporter import generate_html_from_json
+from .queries import special_queries
+from .discord_bot import send_metric_alert
+from .slugs import legacy_asset_slugs
+from .json_processor import create_stable_json
+from .metric_report import MetricReport
+from .slug_report import SlugReport
+from .file_utils import save_json_to_file
+from .s3 import upload_to_s3
+from .config import Config
+from .constants import DATETIME_PATTERN_METRIC, \
+                       HISTOGRAM_METRICS_LIMIT, \
+                       BATCH_SIZE, \
+                       METRICS_WITH_LONGER_DELAY, \
+                       METRICS_WITH_ALLOWED_NEGATIVES, \
+                       INTERVAL_TIMEDELTA, \
+                       ERRORS_IN_ROW, \
+                       PYTHON_ENV
 
 def run(slugs, days_back, interval):
     logging.info('PYTHON_ENV: %s', PYTHON_ENV)
